requirements.txt
numpy
pandas
PyYAML
prometheus-client
